Add instrumentation for graphql 20 that does not use deprecated methods (#10779)
