+++ conventions/src/main/kotlin/io.opentelemetry.instrumentation.base.gradle.kts
@@ -34,16 +34,24 @@ extra["testLatestDeps"] = testLatestDeps
 abstract class TestLatestDepsRule : ComponentMetadataRule {
   override fun execute(context: ComponentMetadataContext) {
     val version = context.details.id.version
-    if (version.contains("-alpha", true) ||
-      version.contains("-beta", true) ||
-      version.contains("-rc", true) ||
-      version.contains("-m", true) || // e.g. spring milestones are published to grails repo
-      version.contains(".alpha", true) || // e.g. netty
-      version.contains(".beta", true) || // e.g. hibernate
-      version.contains(".cr", true) // e.g. hibernate
+    if (version.contains("-alpha", true)
+      || version.contains("-beta", true)
+      || version.contains("-rc", true)
+      || version.contains("-m", true) // e.g. spring milestones are published to grails repo
+      || version.contains(".alpha", true) // e.g. netty
+      || version.contains(".beta", true) // e.g. hibernate
+      || version.contains(".cr", true) // e.g. hibernate
+      || version.endsWith("-nf-execution") // graphql
+      || GIT_SHA_PATTERN.matches(version) // graphql
+      || DATETIME_PATTERN.matches(version) // graphql
     ) {
       context.details.status = "milestone"
     }
+  }
+
+  companion object {
+    private val GIT_SHA_PATTERN = Regex("^.*-[0-9a-f]{7,}$")
+    private val DATETIME_PATTERN = Regex("^\\d{4}-\\d{2}-\\d{2}T\\d{2}-\\d{2}-\\d{2}.*$")
   }
 }
