--- instrumentation/graphql-java-12.0/library/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-  id("otel.library-instrumentation")
-}
-
-dependencies {
-  library("com.graphql-java:graphql-java:12.0")
-
-  testImplementation(project(":instrumentation:graphql-java-12.0:testing"))
-}
