+++ instrumentation/graphql-java/graphql-java-12.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/graphql/v12_0/GraphqlSingletons.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.graphql.v12_0;
+
+import graphql.execution.instrumentation.Instrumentation;
+import io.opentelemetry.api.GlobalOpenTelemetry;
+import io.opentelemetry.instrumentation.graphql.internal.InstrumentationUtil;
+import io.opentelemetry.instrumentation.graphql.v12_0.GraphQLTelemetry;
+import io.opentelemetry.javaagent.bootstrap.internal.InstrumentationConfig;
+
+public final class GraphqlSingletons {
+
+  private static final boolean QUERY_SANITIZATION_ENABLED =
+      InstrumentationConfig.get()
+          .getBoolean("otel.instrumentation.graphql.query-sanitizer.enabled", true);
+
+  private static final GraphQLTelemetry TELEMETRY =
+      GraphQLTelemetry.builder(GlobalOpenTelemetry.get())
+          .setSanitizeQuery(QUERY_SANITIZATION_ENABLED)
+          .build();
+
+  private GraphqlSingletons() {}
+
+  public static Instrumentation addInstrumentation(Instrumentation instrumentation) {
+    Instrumentation ourInstrumentation = TELEMETRY.newInstrumentation();
+    return InstrumentationUtil.addInstrumentation(instrumentation, ourInstrumentation);
+  }
+}
