+++ instrumentation/graphql-java/graphql-java-20.0/library/src/main/java/io/opentelemetry/instrumentation/graphql/v20_0/GraphQLTelemetryBuilder.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.graphql.v20_0;
+
+import com.google.errorprone.annotations.CanIgnoreReturnValue;
+import io.opentelemetry.api.OpenTelemetry;
+
+/** A builder of {@link GraphQLTelemetry}. */
+@SuppressWarnings("AbbreviationAsWordInName")
+public final class GraphQLTelemetryBuilder {
+
+  private final OpenTelemetry openTelemetry;
+
+  private boolean sanitizeQuery = true;
+
+  GraphQLTelemetryBuilder(OpenTelemetry openTelemetry) {
+    this.openTelemetry = openTelemetry;
+  }
+
+  /** Sets whether sensitive information should be removed from queries. Default is {@code true}. */
+  @CanIgnoreReturnValue
+  public GraphQLTelemetryBuilder setSanitizeQuery(boolean sanitizeQuery) {
+    this.sanitizeQuery = sanitizeQuery;
+    return this;
+  }
+
+  /**
+   * Returns a new {@link GraphQLTelemetry} with the settings of this {@link
+   * GraphQLTelemetryBuilder}.
+   */
+  public GraphQLTelemetry build() {
+    return new GraphQLTelemetry(openTelemetry, sanitizeQuery);
+  }
+}
