+++ instrumentation/graphql-java/graphql-java-common/library/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+  id("otel.library-instrumentation")
+}
+
+dependencies {
+  library("com.graphql-java:graphql-java:12.0")
+}
