+++ instrumentation/graphql-java/graphql-java-common/testing/src/main/resources/schema.graphqls
