+++ settings.gradle.kts
@@ -180,9 +180,12 @@ include(":instrumentation:cassandra:cassandra-4.4:library")
 include(":instrumentation:cassandra:cassandra-4.4:testing")
 include(":instrumentation:cassandra:cassandra-4-common:testing")
 include(":instrumentation:cdi-testing")
-include(":instrumentation:graphql-java-12.0:javaagent")
-include(":instrumentation:graphql-java-12.0:library")
-include(":instrumentation:graphql-java-12.0:testing")
+include(":instrumentation:graphql-java:graphql-java-12.0:javaagent")
+include(":instrumentation:graphql-java:graphql-java-12.0:library")
+include(":instrumentation:graphql-java:graphql-java-20.0:javaagent")
+include(":instrumentation:graphql-java:graphql-java-20.0:library")
+include(":instrumentation:graphql-java:graphql-java-common:library")
+include(":instrumentation:graphql-java:graphql-java-common:testing")
 include(":instrumentation:internal:internal-application-logger:bootstrap")
 include(":instrumentation:internal:internal-application-logger:javaagent")
 include(":instrumentation:internal:internal-class-loader:javaagent")
